15 sprint homework. 1. current files removed
