+++ app/src/main/java/com/example/playlistmaker/data/dto/TracksResponse.kt
@@ -1,0 +1,6 @@
+package com.example.playlistmaker.data.dto
+
+
+import com.example.playlistmaker.domain.models.Track
+
+class TracksResponse(val resultCounts:Int, val results:List<Track>)
