+++ app/src/main/java/com/example/playlistmaker/data/network/ItunesApi.kt
@@ -1,5 +1,6 @@
-package com.example.playlistmaker.netconnection
+package com.example.playlistmaker.data.network
 
+import com.example.playlistmaker.data.dto.TracksResponse
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
