--- app/src/main/java/com/example/playlistmaker/netconnection/TracksResponse.kt
@@ -1,5 +1,0 @@
-package com.example.playlistmaker.netconnection
-
-import com.example.playlistmaker.Track
-
-class TracksResponse(val resultCounts:Int, val results:List<Track>)
