+++ app/src/main/java/com/example/playlistmaker/ui/main/MainActivity.kt
@@ -1,10 +1,14 @@
-package com.example.playlistmaker
+package com.example.playlistmaker.ui.main
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.View
 import android.widget.Button
+import com.example.playlistmaker.ui.media.MediaActivity
+import com.example.playlistmaker.R
+import com.example.playlistmaker.ui.settings.SettingActivity
+import com.example.playlistmaker.ui.search.SearchActivity
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -16,17 +20,17 @@ class MainActivity : AppCompatActivity() {
         val settingButton = findViewById<Button>(R.id.settingsButton)
 
         searchButton.setOnClickListener {
-            val intent = Intent(this@MainActivity,SearchActivity::class.java)
+            val intent = Intent(this@MainActivity, SearchActivity::class.java)
             startActivity(intent)
         }
         val mediaButtonClickListener = View.OnClickListener {
-            val intent = Intent(this@MainActivity,MediaActivity::class.java)
+            val intent = Intent(this@MainActivity, MediaActivity::class.java)
             startActivity(intent)
         }
         mediaButton.setOnClickListener(mediaButtonClickListener)
 
         settingButton.setOnClickListener{
-            val intent = Intent(this@MainActivity,SettingActivity::class.java)
+            val intent = Intent(this@MainActivity, SettingActivity::class.java)
             startActivity(intent)
         }
     }
