+++ app/src/main/java/com/example/playlistmaker/ui/media/MediaActivity.kt
@@ -1,7 +1,8 @@
-package com.example.playlistmaker
+package com.example.playlistmaker.ui.media
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.playlistmaker.R
 
 class MediaActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
