+++ app/src/main/java/com/example/playlistmaker/ui/search/SearchViewHolder.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.searchrecycler
+package com.example.playlistmaker.ui.search
 
 import android.content.Context
 import android.util.TypedValue
@@ -9,7 +9,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
 import com.example.playlistmaker.R
-import com.example.playlistmaker.Track
+import com.example.playlistmaker.domain.models.Track
 import java.text.SimpleDateFormat
 import java.util.Locale
